auth: accept LSAT/L402 in auth headers
Add support for both "LSAT <macaroon>:<preimage>" and "L402 <macaroon>:<preimage>" when
validating an auth header.

The `?:` in the regular expression creates a non-capturing group for "LSAT" or "L402".
This means that we match either "LSAT" or "L402", but we don't capture this match separately.
This is necessary to ensure that `FindStringSubmatch` returns the expected number of matches so
the changes in the code are minimal. Without the non-capturing group, it would return an extra
match for the "LSAT" or "L402" part of the regular expression.

 Changes to be committed: